requirements.txt
# Web框架
fastapi==0.104.1
uvicorn==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1

# 机器学习/AI
openai==1.3.7
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# 文档处理
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1

# 向量数据库
chromadb==0.4.18
qdrant-client==1.6.9

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 工具库
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# 异步任务
aioredis==2.0.1
celery==5.3.4

# 监控和日志
prometheus-client==0.19.0
structlog==23.2.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1